features/quota: use signed type to store delta.

BUG: 2604 (Quota: crossing the set limit)
URL: http://bugs.gluster.com/cgi-bin/bugzilla3/show_bug.cgi?id=2604

+++ xlators/features/quota/src/quota.h
@@ -148,7 +148,7 @@ struct quota_local {
         loc_t        oldloc;
         loc_t        newloc;
         loc_t        validate_loc;
-        uint32_t     delta;
+        int64_t      delta;
         int32_t      op_ret;
         int32_t      op_errno;
         int64_t      size;
